hrms/requirements.txt
Django>=6.0.0,<7.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
django-cors-headers>=4.3.0

